Orders Dashboard: CoffeeScript Updated
Updated the CoffeeScripts for Orders Dashboard. Fixed the problem with "Confirm Submission" message showing up when deleting one or more orders from the screen. This was caused because of the underlying AJAX call. Duplicated the way jquery_ujs manages to do this without raising the confirmation dialog box. Basically changed the type of call from DELETE to POST and added the hidden field _method=delete to designate that this is a delete operation. However CSRF has not been addressed during this operation and will need to be addressed at a later date.

Add AXLSX gem to the project to start testing out the feasibility of creating Excel files that do not popup the warning in Excel about the file format not matching the extension. Also this should help with opening files in Google Docs and OpenOffice. Thus making the file accessible on any machine.