Rails 4 Upgrade
Upgrading the environments/development.rb for the Rails 4 upgrade. Also removed references to attr_accessible from all models and replaced it with params[:model].permit in the controllers. Need to find a better way to define these private methods for these models to be used in controllers where these models are accessed.