Elasticsearch Implementation (In Progress)
Have worked out the problems with indexing and mapping. However, cannot get mult_field mapping working properly in the mapping. After changes to the mapping and the as_indexed_json method the index is not being populated at all. Console is not producing any errors. Store name with locality is the field that requires multi_field functionality.

Although basic mapping has been worked out the proper analyzers will still need to be applied for searching to be efficient.

 On branch r417upgrade
 Changes to be committed:
	modified:   Gemfile
	modified:   app/models/concerns/store_searchable.rb
	modified:   app/models/store.rb
	new file:   app/modules/store_import.rb