r417Upgrade - Store Search - IN PROGRESS
Store search is being slowly implemented. At this time the store listing by company and states is working. However the AJAX call for stores by states within the company VIEW page is not working.

Also the store listing page does not have all the details for the stores at this time. The _store.html.erb will need to be updated to provide all the information that is currently available in the live site.

 Changes to be committed:
	modified:   app/models/concerns/store_searchable.rb
	modified:   app/views/stores/_stores.html.erb