r417 Upgrade: Stores > Facet Updates
Transitioning from Facets to Aggregations as part of the updates. Facets in ES are due to be deprecated in the future. Still sorting out some of the issues with Aggregations. The primary issue is not being able to successfully sort by the REGION_NAME. At the moment this can be handled on the Ruby end by sorting it using Ruby functions instead of having ES return a sorted aggregation key list.

This update also includes an update to the mapping so it will need to be merged with the branch that has been created for fine tuning the mapping and searching mechanism.

 Changes to be committed:
	modified:   app/models/concerns/store_searchable.rb