Store > Add > Updates to Regions
The primary focus of this commit was to address the scenarios revolving around the Regions dropdown. There is just one scenario left that will make sure there is a way to add new regions right from the new store entry page. Also if a region is no longer being referenced by any store it should be removed from the database. This will eliminate the need for elaborate CRUD screens for Regions as well.

To address the Regions dropdown, the idea of APIs was explored and as a result the following changes were made:

- Created new directory structure within Controllers to organize the controllers for APIs and version them as well. This was done more to facilitate proper routing than for organization.
- Routing was updated so that referencing the APIs will be easier. Eventually all actions that provide data to AJAX calls will be moved into the API controllers.

 Changes to be committed:
	modified:   Gemfile
	modified:   app/assets/javascripts/google_maps.js.coffee
	new file:   app/assets/javascripts/store_new_edit.js.coffee
	new file:   app/controllers/api/api_controller.rb
	new file:   app/controllers/api/v1/api_controller.rb
	new file:   app/controllers/api/v1/regions_controller.rb
	modified:   app/controllers/application_controller.rb
	modified:   app/models/store.rb
	modified:   app/views/stores/new.html.erb
	new file:   config/initializers/geocoder.rb
	deleted:    config/initializers/geokit_config.rb
	modified:   config/routes.rb