Store > Add & Update > Duplicate Check - Basic
Very basic UI implemented for duplicate search. The UI will need to be tweaked and tested for update as well. Update validation for duplicates has not been applied as of yet. There is a bug at the moment which I am trying to resolve. If a user were to copy and paste an address in the address field without selecting one of the address that Google provides, the various hidden address componenets are populated/updated. To correct this problem, have to make sure that the "blur" event triggers a Google lookup and populates the address components as needed. What I don't know right now is if I can use the "getPlace()" function in the Google library against an address string instead of the input element.

General cleanup of gems that are no longer needed for this project. Also removing images that are no longer used.