Audit > Updating ES and Model Config
Inital changes to adapt the model for the upcoming audit sheet changes. At the moment the model associations are not working properly between audit, audit_metric and audit_metric_responses. This is possibly because audit_metric and audit_metric_responses do not have `id` field. So primary key configuration has to be configured properly for associations to work.

Also made changes for ES to work properly. These are untested at the time of commit.