Audits > Association Fix & ES Refinements
Fixed issues with associations not properly working. Also fixed the mapping for Audits and revised the Store mapping as well to store more information about the last_audit. Also forced the Store document to be re-indexed every time there is a change to an associated Audit. This will need further refinement to make sure that the the Store document is only indexed when the latest Audit is modified.