Audits > Add New - in progress
This may be the final database structure for the audits section. Some notes about this design:

It has been concluded that the logic for computing the score will reside in the view using Angular. This logic will not be stored in the database. This will definitely cause problems in the future. In the event that audit sheet is changed and older audits need to be pulled, it may create problems in presenting the data that is historically accurate because the computational logic would have changed in the new view. So view versioning may have to be looked into.