Audit Updates > Add New Audit > Ready made Image Upload (AngularJS)
Using ready made AngularJS image upload functionality with S3. This is not a preferred solution as the code produces a lot of the HTML that is displayed for user input. There are other lightweight options that can be employed. At the moment the following solution is being employed:

https://github.com/asafdav/ng-s3upload

Next iteration will be to try this solution:
http://www.cheynewallace.com/uploading-to-s3-with-angularjs/
https://github.com/cheynewallace/angular-s3-upload/blob/master/index.html

Uploading it directly will ensure that the web server is not unnecessarily stressed with the job of accepting the file and then uploading it to the CDN. Amazon S3 has good bandwidth and speed. It can handle the direct upload from the client browser. However, this creates the following issues:

- Files are not completely validated to make sure that users are not uploading system executable files masked as image files. For example a user can change the file extension of an executable file to a JPEG and upload it. Javascript is only able to validate the extension of the file and not the type.

- In case an audit is abandoned midway with an image attached, this will create an orphan image in the bucket.