Audit Updates > Add New Audit > Issue
Discovered that the save function is not actually writing records to the the audit_metric_responses table at all. Everything seems to be setup correctly but it is just not work.

What is working though is the part that writes the image URL to the database. Next part of that is to make a background service that makes sure that the images are moved from the direct_uploads folder to another folder. The S3 bucket lifecycle will delete any images in that folder that are older than a day.