Audit Updates > Add New Audit > Audit Metric Response (Resolved)
Finally resolved the issue with the AuditMetricResponse not being populated on Audit.create. The issue was very simple - the audit_metric_responses_attributes always needed to be present otherwise the deserializer would ignore them. The deserializers looks at the first entry coming in to determine the rest of the data. Since the first results did not contain the audit_metric_responses_attributes it would go ahead and ignore it for the rest of the entries as well. Right now a patch has been applied in the way of passing a dummy audit_metric_response attribute for the type of entries that will not contain one and the metric_option_id for such attributes has been set to 0. In the model, the accepts_nested_attributes_for has been given an extra condition to reject any nested attributes that have the metric_option_id set to 0.

To make the code more consistent and less hacky, the audit_metric_responses attributes will be included with the radio button entries as well.